@@ -36,15 +36,17 @@ import net.tridentsdk.util.Vector;
  */
 @InternalUseOnly
 public class EntityTracker {
-    public void track(Entity entity, byte... meta) {
+    public void track(Entity entity) {
         if (entity instanceof TridentPlayer)
             return;
+
         PacketPlayOutSpawnMob packet = new PacketPlayOutSpawnMob();
+
         packet.set("entityId", entity.entityId())
                 .set("type", entity.type())
                 .set("entity", entity)
                 .set("metadata", ((TridentEntity) entity).protocolMeta);
-        // TODO
+
         TridentPlayer.sendAll(packet);
         ((TridentWorld) entity.world()).addEntity(entity);
     }
@@ -67,11 +69,13 @@ public class EntityTracker {
 
     private void sendMove(Entity entity, Position to, Vector diff) {
         PacketPlayOutEntityCompleteMove move = new PacketPlayOutEntityCompleteMove();
+
         move.set("entityId", entity.entityId())
                 .set("difference", diff)
                 .set("yaw", to.yaw())
                 .set("pitch", to.pitch())
                 .set("flags", (byte) 0);
+
         TridentPlayer.sendAll(move);
     }
 }
