@@ -17,13 +17,11 @@
 
 package net.tridentsdk.server.window;
 
-import com.google.common.base.Function;
 import com.google.common.collect.Collections2;
 import io.netty.util.internal.chmv8.ConcurrentHashMapV8;
 import net.tridentsdk.window.Window;
 import net.tridentsdk.window.WindowHandler;
 
-import javax.annotation.Nullable;
 import javax.annotation.concurrent.ThreadSafe;
 import java.util.Collection;
 import java.util.Map;
@@ -49,11 +47,6 @@ public class TridentWindowHandler implements WindowHandler {
 
     @Override
     public Collection<Window> windows() {
-        return Collections2.transform(windows.values(), new Function<TridentWindow, Window>() {
-            @Nullable @Override
-            public Window apply(TridentWindow tridentWindow) {
-                return tridentWindow;
-            }
-        });
+        return Collections2.transform(windows.values(), (w) -> w);
     }
 }
